+++ src/index.js
@@ -5,10 +5,27 @@ import quiltrosRouter from "./routes/quiltros.js";
 import eventsRouter from "./routes/events.js";
 import s3Router from "./routes/s3.js";
 import functions from "firebase-functions";
+import dotenv from "dotenv"
+import mongoose from "mongoose";
 
 const PORT = 3000;
 
+dotenv.config();
+
 export let dbClient;
+
+const connectDb = async () => {
+  try {
+    const mongoUrl = process.env.MONGO_DB_URL;
+    dbClient = await mongoose.connect(mongoUrl, { dbName: "quiltros" });
+    console.log("DATABASE CONNECTED, APP STARTING");
+    return dbClient;
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+connectDb();
 
 const app = express();
